requirements.txt
flask
mysql-connector-python
pillow
tkcalendar